refactor(crates): split FUSE machinery into heph-driver-bridge so plugins stay fuse-free
Fuse linkage flowed through driver-support -> sandboxfuse, and every plugin
depends on driver-support to implement the ManagedDriver trait. So every plugin
transitively linked libfuse and had to emit the -weak-lfuse link arg via its own
build.rs.

Split driver-support:

- driver-support (fuse-free contract): ManagedDriver trait, ManagedRun* types,
  ShellFallback, ManagedDriverOs (OS-copy sandbox), FuseMode enum, shared
  output/source-map helpers. Dropped the sandboxfuse dep and build.rs; the
  shared helpers + ManagedDriverOs are now pub.
- driver-bridge (new, links fuse): ManagedDriverBridge (OS/FUSE router),
  FuseSlot, ManagedDriverFuse. Keeps the -weak-lfuse build.rs. Only the engine
  depends on it (engine already linked sandboxfuse directly to mount the FS).

plugin-exec's bridge test retargets to ManagedDriverOs::run_inner (the OS path
the bridge dispatches to with FUSE off), keeping the test fuse-free.

Result: plugin-exec, plugin-go, plugin-nix, plugin-stabby, plugin-sdk and
plugin-go-cdylib drop their build.rs and no longer link fuse. cargo tree -i
sandboxfuse now shows only driver-bridge + engine. Plugins know only the
Driver/ManagedDriver interface, nothing about FUSE.
