+++ crates/driver-support/Cargo.toml
@@ -10,7 +10,6 @@ workspace = true
 hcore = { package = "core", path = "../core" }
 hmodel = { package = "model", path = "../model" }
 hplugin = { package = "plugin", path = "../plugin" }
-hsandboxfuse = { package = "sandboxfuse", path = "../sandboxfuse", default-features = false }
 anyhow = "1.0.102"
 async-trait = "0.1.89"
 serde = { version = "1", features = ["derive"] }
@@ -18,7 +17,6 @@ serde_json = "1.0"
 tar = "0.4"
 glob = "0.3"
 walkdir = "2"
-rayon = "1.12.0"
 xxhash-rust = { version = "0.8.15", features = ["xxh3", "std"] }
 tokio = { version = "1.52.3", features = ["rt", "io-util", "fs", "process"] }
