+++ crates/engine/Cargo.toml
@@ -15,6 +15,7 @@ hlock = { package = "lock", path = "../lock" }
 hsandboxfuse = { package = "sandboxfuse", path = "../sandboxfuse", default-features = false }
 hplugin = { package = "plugin", path = "../plugin" }
 hdriver-support = { package = "driver-support", path = "../driver-support" }
+hdriver-bridge = { package = "driver-bridge", path = "../driver-bridge" }
 hbuiltins = { package = "builtins", path = "../builtins" }
 hplugin-exec = { package = "plugin-exec", path = "../plugin-exec" }
 hplugin-query = { package = "plugin-query", path = "../plugin-query" }
